Update expertiseAccoladesPhrases.js
These are completed, and with 4 sentence variants per 6 accolade results for each of the 8 fields of expertise, I think that means a 1/192 chance that 2 characters will have the same readout, and a 1/24 chance that two of the same field will (though this probability is altered by the 4d6 roll that draws most rolls toward the "one equaled pair" option). Still, it's a good mix of variation I think!